Added some blind support
Ok so basically, the player now has a directional light source that can be toggled on and off. To simulate the blind effect, disable all other lights, enable that one, and take the "Blind" material and throw that shit everywhere. The effect is all materials are completely black but light reflects off of them. Still working on making the light's FOV more narrow.